+++ CHANGELOG.md
@@ -27,6 +27,10 @@ Other blocks are:
     * `broaden` expects a function (or `None`) to include contributions from x into the definition of y.
     * The internal *qlabel* has changed: the old sub-option is now divided into a quantity-specific option and the level of theory (`H`, `A`).
 
+### Added
+- **LIB** - New function `convert_expr` in `tools/char` to convert simple "human" mathematical expressions to Python-valid ones.
+- **LIB** - New submodule `char` in `tools/` for string-related operations.
+
 ### Fixed
 - **API** - `parse_qlabel` correctly parses the sub-option of `qty_tag=2` (same as `atcrd`).
 - **PY** - Changed all import aliases inside ESTAMPES to have the form `from xxx import yyy as zzz`.  This should reduce the number of problems with *PIP* installations.